+++ lib/linux/index.js
@@ -1,0 +1,2 @@
+
+throw new Error('Platform linux is unsupported for now!');
