+++ lib/util.js
@@ -47,7 +47,7 @@ exports.parseSettings = function(value) {
     proxyEnable: false,
     proxyServer: '',
     bypassLocal: false,
-    bypass: '',
+    bypass: ''
   };
   if (!value || !value.length) {
     return result;
